+++ CHANGELOG.md
@@ -1,6 +1,12 @@
+## 1.0.4 - 2018-12-10
+### Fixed
+- Fixed saving new reverse relations
+- Fixed bug when importing with Schematic when the target field didn't yet exist
+- Fixed bug where targets where validated
+
 ## 1.0.3 - 2018-12-07
 ### Fixed
-- Fixed bug when importing in Schematic
+- Fixed bug when importing with Schematic
 
 ## 1.0.2 - 2018-12-07
 ### Added